cn_project/requirements.txt
pycryptodome==3.9.9
py7zr==0.11.0